VO 수정

+++ src/main/java/com/teamproject/trackers/biz/comment/CommentVO.java
@@ -3,22 +3,22 @@ package com.teamproject.trackers.biz.comment;
 import java.sql.Date;
 
 public class CommentVO {
-	private int post_id;
-	private int id;
+	private long post_id;
+	private long id;
 	private String content;
 	private Date cre_date;
 	
 	
-	public int getPost_id() {
+	public long getPost_id() {
 		return post_id;
 	}
-	public void setPost_id(int post_id) {
+	public void setPost_id(long post_id) {
 		this.post_id = post_id;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getContent() {
