+++ src/main/java/com/teamproject/trackers/biz/followSubscribeLike/SubscribeVO.java
@@ -3,21 +3,21 @@ package com.teamproject.trackers.biz.followSubscribeLike;
 import java.sql.Date;
 
 public class SubscribeVO {
-	private int subscibe_id;
-	private int id;
+	private long subscibe_id;
+	private long id;
 	private Date subscibe_date;
 	
 	
-	public int getSubscibe_id() {
+	public long getSubscibe_id() {
 		return subscibe_id;
 	}
-	public void setSubscibe_id(int subscibe_id) {
+	public void setSubscibe_id(long subscibe_id) {
 		this.subscibe_id = subscibe_id;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public Date getSubscibe_date() {
